requirements.txt
# Generated from pyproject.toml
# Python version: >=3.11

flask>=3.1.2
flask-sqlalchemy>=3.1.1
sqlalchemy>=2.0.43
werkzeug>=3.1.3
flask-login>=0.6.3
email-validator>=2.3.0

# Integrations / services
twilio>=9.8.0
slack-sdk>=3.36.0
sendgrid>=6.12.4
notion-client>=2.5.0

# Content and parsing
trafilatura>=2.0.0

# Validation / models
pydantic>=2.11.7

# Google GenAI SDK
google-genai>=1.33.0

# Server / deployment
gunicorn>=23.0.0

# Database drivers (binary wheels)
psycopg2-binary>=2.9.10

# Misc utilities
sift-stack-py>=0.8.5